+++ src/containers/EntityTable/columns/__tests__/name.test.js
@@ -1,0 +1,16 @@
+/*
+ * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
+ */
+
+import { nameColumn } from '../name';
+
+describe('nameColumn', () => {
+  it('is configured correctly', () => {
+    expect(nameColumn).toMatchSnapshot();
+  });
+  describe('Cell', () => {
+    it('renders correctly', () => {
+      expect(nameColumn.Cell({ row: { name: 'mock name' } })).toMatchSnapshot();
+    });
+  });
+});