+++ src/containers/Form/Lists/validation.js
@@ -2,14 +2,11 @@
  * Copyright © 2015-2017 Serenova, LLC. All rights reserved.
  */
 
-import { atLeastOneLetterOrNumber } from '../../../utils/string';
+import { isEmpty } from '../../../utils/string';
 
 export const updateFormValidation = values => {
   const errors = {};
-  if (
-    !values.get('name') ||
-    !atLeastOneLetterOrNumber.test(values.get('name'))
-  ) {
+  if (isEmpty(values.get('name'))) {
     errors.name = 'Please enter a name';
   }
   return errors;