+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.19.5]
+* CXV1-13304 - Validation, string, and Entity table config unit tests
+* CXV1-13603 - Add email templates help doc link
+* Fix entity side panel details
+
 ## [0.19.4]
 * CXV1-13304 - Epic Unit Tests